requirements.txt
openai-whisper>=20231117
transformers>=4.35.0
torch>=2.0.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-docx>=0.8.11
Pillow>=10.0.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0
